Rendering process moved to a different thread
This is *much* better now. It should be easy to wait for other events
in the main thread without turning the GUI unresponsive.

With the rendering occurring in a different thread now is possible to
use the SDL_WaitEvent() instead of SDL_PollEvent() so the CPU and GPU
usage is much lower.